"Rename Layout folder for Dashboard component files
